Podcast-summary/requirements.txt
streamlit
google-generativeai
pydub
SpeechRecognition
ffmpeg-python
